requirements.txt
python-dotenv==1.0.0
supabase==2.0.3
openai==1.3.0
Jinja2==3.1.2
fastapi==0.104.1
uvicorn==0.24.0.post1
python-multipart==0.0.6
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
Pillow==9.5.0